add to setup.py + specify int64 in test

+++ pandas/tests/internals/test_external_block.py
@@ -14,7 +14,7 @@ class CustomBlock(Block):
 
 
 def test_custom_repr():
-    values = np.arange(3)
+    values = np.arange(3, dtype='int64')
 
     # series
     block = CustomBlock(values, placement=slice(0, 3))
